// web/src/api/generated/models/TransferExportDetail.ts
export interface TransferExportDetail {
    /**
     * 
     * @type {number}
     * @memberof TransferExportDetail
     */
    artifactSize: number | null;
    /**
     * 
     * @type {Date}
     * @memberof TransferExportDetail
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof TransferExportDetail
     */
    downloadUrl: string | null;
    /**
     * 
     * @type {string}
     * @memberof TransferExportDetail
     */
    errorCode: string | null;
    /**
     * 
     * @type {Date}
     * @memberof TransferExportDetail
     */
    expiresAt: Date;
    /**
     * 
     * @type {string}
     * @memberof TransferExportDetail
     */
    id: string;
    /**
     * 
     * @type {Date}
     * @memberof TransferExportDetail
     */
    readyAt: Date | null;
    /**
     * 
     * @type {TransferScope}
     * @memberof TransferExportDetail
     */
    scope: TransferScope;
    /**
     * 
     * @type {ExportStatus}
     * @memberof TransferExportDetail
     */
    status: ExportStatus;
}