// web/src/api/generated/models/TransferImportDetail.ts
export interface TransferImportDetail {
    /**
     * 
     * @type {number}
     * @memberof TransferImportDetail
     */
    artifactSize: number;
    /**
     * 
     * @type {Date}
     * @memberof TransferImportDetail
     */
    completedAt: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof TransferImportDetail
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof TransferImportDetail
     */
    errorCode: string | null;
    /**
     * 
     * @type {Date}
     * @memberof TransferImportDetail
     */
    expiresAt: Date;
    /**
     * 
     * @type {string}
     * @memberof TransferImportDetail
     */
    id: string;
    /**
     * 
     * @type {TransferScope}
     * @memberof TransferImportDetail
     */
    scope: TransferScope | null;
    /**
     * 
     * @type {ImportStatus}
     * @memberof TransferImportDetail
     */
    status: ImportStatus;
    /**
     * 
     * @type {TransferImportSummary}
     * @memberof TransferImportDetail
     */
    summary: TransferImportSummary | null;
    /**
     * 
     * @type {Date}
     * @memberof TransferImportDetail
     */
    validatedAt: Date | null;
}