// web/src/api/generated/models/ImportStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SideBySide Next
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ImportStatus = {
    QUEUED: 'QUEUED',
    VALIDATING: 'VALIDATING',
    READY_TO_APPLY: 'READY_TO_APPLY',
    APPLYING: 'APPLYING',
    COMPLETED: 'COMPLETED',
    FAILED: 'FAILED',
    EXPIRED: 'EXPIRED'
} as const;
export type ImportStatus = typeof ImportStatus[keyof typeof ImportStatus];


export function instanceOfImportStatus(value: any): boolean {
    for (const key in ImportStatus) {
        if (Object.prototype.hasOwnProperty.call(ImportStatus, key)) {
            if (ImportStatus[key as keyof typeof ImportStatus] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ImportStatusFromJSON(json: any): ImportStatus {
    return ImportStatusFromJSONTyped(json, false);
}

export function ImportStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): ImportStatus {
    return json as ImportStatus;
}

export function ImportStatusToJSON(value?: ImportStatus | null): any {
    return value as any;
}

export function ImportStatusToJSONTyped(value: any, ignoreDiscriminator: boolean): ImportStatus {
    return value as ImportStatus;
}